Secured Eval, Updated Daily
Eval now is for super admins only. Daily now displays amount of time until you case use it again.

+++ commands/eval.js
@@ -7,14 +7,25 @@ exports.helpText = "Runs the code. Meant for debugging only.";
 
 //Requirements
 var main = require('../main.js');
+var config = require('../config.json');
 const Discord = require('Discord.js');
+
+function isSuperAdmin(id){
+    if (config.adminids.includes(id))
+            return true;
+        return false;
+}
 
 //Command
 exports.command = async function (message, args, temptotals, exp) {
+    if (!isSuperAdmin(message.author.id))
+        return message.reply("You don't have permissions to do that.");
     try{
-        eval(args.join(''));
+        var code=args.join('');
+        eval(code);
     } catch (err) {
-        message.author.send(err);
+        console.log("Eval Err");
+        message.author.send("Error"+err);
     }
     return await message.delete();
 }
